Updated tests to work with the newest version of scipy (1.5)
Scipy 1.5 release notes:

The wilcoxon function now computes the p-value for Wilcoxon’s signed rank test using the exact distribution for inputs up to length 25. The function has a new mode parameter to specify how the p-value is to be computed. The default is "auto", which uses the exact distribution for inputs up to length 25 and the normal approximation for larger inputs.